rafactor: レビュー部分をコンポーネント化
